Revert "cfg80211: Add support for 60GHz band channels 5 and 6"
This reverts commit 7a9151c900762f4dee678c273607c2a7b0826dab.

Bug: 185310516

+++ drivers/net/wireless/ath/wil6210/debugfs.c
@@ -1598,7 +1598,7 @@ static int wil_freq_debugfs_show(struct seq_file *s, void *data)
 {
 	struct wil6210_priv *wil = s->private;
 	struct wireless_dev *wdev = wil->main_ndev->ieee80211_ptr;
-	u32 freq = wdev->chandef.chan ? wdev->chandef.chan->center_freq : 0;
+	u16 freq = wdev->chandef.chan ? wdev->chandef.chan->center_freq : 0;
 
 	seq_printf(s, "Freq = %d\n", freq);
